camslide_update: f32 approach local drops spurious frsp, fixes store order (98.49->98.75)

+++ src/main/dll/dll_0042_unk.c
@@ -244,7 +244,7 @@ void camslide_update(CameraObject* camera, GameObject* target, f32 upperBound, f
     f32 low;
     f32 range;
     f32 slide;
-    f64 approach;
+    f32 approach;
     f32 mtx[16];
     CamSlideRot rot;
     f32 velX;
